replace useFont with fontConfig theme provider property
