+++ src/docs/Wrapper.js
@@ -1,7 +1,7 @@
 import React from "react";
 import PropTypes from "prop-types";
 
-import { ThemeProvider, defaultLightTheme, useFont } from "../index";
+import { ThemeProvider, defaultLightTheme } from "../index";
 import SourceSansProLight from "../assets/fonts/source-sans-pro/SourceSansPro-Light.ttf";
 import SourceSansProRegular from "../assets/fonts/source-sans-pro/SourceSansPro-Regular.ttf";
 import SourceSansProSemiBold from "../assets/fonts/source-sans-pro/SourceSansPro-SemiBold.ttf";
@@ -15,11 +15,11 @@ const fontConfig = {
 };
 
 const DoczWrapper = ({ children }) => {
-  // useFont(fontConfig);
   return (
     <ThemeProvider
       themes={{ light: defaultLightTheme }}
-      defaultThemeName="light">
+      defaultThemeName="light"
+      fontConfig={fontConfig}>
       {children}
     </ThemeProvider>
   );
