+++ src/theme/ThemeProvider.js
@@ -1,20 +1,28 @@
-import React, { useState, useCallback, useLayoutEffect } from "react";
+import React, { useState, useCallback, useLayoutEffect, useMemo } from "react";
 import PropTypes from "prop-types";
 
 export const ThemeContext = React.createContext();
 
-export const ThemeProvider = ({ themes, defaultThemeName, children }) => {
+export const ThemeProvider = ({
+  themes,
+  defaultThemeName,
+  fontConfig,
+  children
+}) => {
   const [themeName, setThemeName] = useState(defaultThemeName);
   const updateTheme = useCallback((newThemeName) => {
     setThemeName(newThemeName);
   });
-
+  const currentTheme = useMemo(() => themes && themeName && themes[themeName]);
   useLayoutEffect(() => {
-    if (themeName & themes) {
+    if (currentTheme) {
       const cssVars = themes[themeName];
       Object.keys(cssVars).forEach((key) => {
         document.documentElement.style.setProperty(`--${key}`, cssVars[key]);
       });
+      if (fontConfig) {
+        applyFonyAssets(fontConfig, currentTheme);
+      }
     }
   }, [themeName, themes]);
 
@@ -33,5 +41,41 @@ export const ThemeProvider = ({ themes, defaultThemeName, children }) => {
 ThemeProvider.propTypes = {
   themes: PropTypes.object.isRequired,
   defaultThemeName: PropTypes.string.isRequired,
+  fontConfig: PropTypes.string.object,
   children: PropTypes.node
+};
+
+const applyFonyAssets = (fontConfig, theme) => {
+  var newStyle = document.createElement("style");
+  newStyle.appendChild(
+    document.createTextNode(`
+@font-face {
+  font-family: ${fontConfig.fontFamilyText};
+  src: url(${fontConfig.semiBoldUrl})
+    format("${fontConfig.format}");
+  font-weight: ${theme["font-weight-semi-bold"]};
+  font-style: normal;
+}
+@font-face {
+  font-family: ${fontConfig.fontFamilyText};
+  src: url(${fontConfig.lightUrl})
+    format("${fontConfig.format}");
+  font-weight: ${theme["font-weight-light"]};
+  font-style: normal;
+}
+@font-face {
+  font-family: ${fontConfig.fontFamilyText};
+  src: url(${fontConfig.regularUrl})
+    format("${fontConfig.format}");
+  font-weight: ${theme["font-weight-regular"]};
+  font-style: normal;
+}
+`)
+  );
+  newStyle.type = "text/css";
+  document.head.appendChild(newStyle);
+  document.documentElement.style.setProperty(
+    "--font-family-text",
+    fontConfig.fontFamilyText
+  );
 };
