+++ src/theme/index.js
@@ -1,5 +1,4 @@
 export { default as defaultDarkTheme } from "./darkTheme";
 export { default as defaultLightTheme } from "./lightTheme";
-export { default as useFont } from "./useFont";
 export * from "./ThemeProvider";
 export * from "./helpers";
